RemoteExecutionService: download stdout/err conditionally w. shouldDownload
This is a patch that demonstrate how we could leverage shouldDownload on
the Action's output files and output directories to determine whether we
should download the stdout and stderr of the action. Keep in mind that
`shouldDownload()` includes the following logics underneath:
- Check for Bazel output service (FUSE) usage
- Check on exit code == 0
- Check on --remote_download_outputs flag
- Check on top-level target
- Check on download regexes

This kinda work, but some tests under
`//src/test/shell/bazel/remote:remote_execution_test` are failing
unexpectedly and would require a bit more investigation. It's a bit
concerning to me that tests using disk_cache are failing with this
patch.

There is also an obvious problem with test actions:
While all valid build actions do include `some` mandatory output files
and/or output directories, test actions do NOT.
Test actions do include _optional_ outputs but don't mandate them to
exists, which makes sense.
With this change, we will cause successful test actions that does not
produce any output files/directories to never download their
stderr/stdout. This could be a sensible choice for some setups, because
such a test is uninteresting and not actionable. However, it could still
be a surprise to some folks.

This made me think that we would need another flag to guard this???