GrpcCacheClient: add idle timeout to bytestream read
Currently for grpc remote cache blob download through ByteStream.Read
RPC calls, we enforce a total duration timeout via `--remote_timeout`.
With the change in https://github.com/bazelbuild/bazel/pull/26267, we
would still be enforcing the same total duration timeout, but with a
higher value.

In addition to that, the user can enable `--grpc_keepalive_time`,
together with `--grpc_keepalive_timeout` to send ping frames in the
underlying http2 connection to check for networking issue.

There remains a failure mode where the server is composed of a load
balancer and a server backend. In that case, the load balancer could be
setup to ack the ping and keep the http2 connection alive, but the
server backend could hang with no response messages send to Bazel. In
such a scenario, Bazel may hang until the long total duration timeout is
reached before issuing any retry.

Add an new interceptor to ByteStream.Read calls so that we can check for
idle response stream. This interceptor shall use the shorter timeout
value from `--remote_timeout` flag and trigger failure earlier so that
Bazel may retry the download.